+++ pkgimage.mk
@@ -65,6 +65,7 @@ $(eval $(call sysimg_builder,SHA,))
 $(eval $(call sysimg_builder,Serialization,))
 $(eval $(call sysimg_builder,Sockets,))
 $(eval $(call sysimg_builder,Unicode,))
+$(eval $(call pkgimg_builder,Profile,))
 
 # 1-depth packages
 $(eval $(call pkgimg_builder,GMP_jll,Artifacts Libdl))
@@ -96,9 +97,7 @@ $(eval $(call sysimg_builder,Distributed,Random Serialization Sockets))
 $(eval $(call sysimg_builder,Future,Random))
 $(eval $(call sysimg_builder,InteractiveUtils,Markdown))
 $(eval $(call sysimg_builder,LibGit2,NetworkOptions Printf SHA Base64))
-$(eval $(call sysimg_builder,Profile,Printf))
 $(eval $(call sysimg_builder,UUIDs,Random SHA))
-
 
  # 3-depth packages
  # LibGit2_jll
